Move opeartors into opeartors directory
